+++ PsychShift/Assets/Code Department/Scripts/IDamageable.cs
@@ -2,16 +2,16 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public int CurrentHealth { get; }
-    public int MaxHealth { get; }
+    public float CurrentHealth { get; }
+    public float MaxHealth { get; }
 
     public bool IsWeakPoint { get; }
 
-    public delegate void TakeDamageEvent(int Damage);
+    public delegate void TakeDamageEvent(float Damage);
     public event TakeDamageEvent OnTakeDamage;
 
     public delegate void DeathEvent(Transform Position);
     public event DeathEvent OnDeath;
 
-    public void TakeDamage(int Damage, Guns.GunType gunType);
+    public void TakeDamage(float Damage, Guns.GunType gunType);
 }
