+++ PsychShift/Assets/Code Department/Scripts/KevinScripts/PuzzleKit/Split puzzle kit/ActionBox.cs
@@ -32,10 +32,10 @@ public class ActionBox : MonoBehaviour, IDamageable
     public event IDamageable.DeathEvent OnDeath;
     #pragma warning restore 67
 
-    private int currentHealth;
-    public int CurrentHealth { get => currentHealth; set => currentHealth = value;}
+    private float currentHealth;
+    public float CurrentHealth { get => currentHealth; set => currentHealth = value;}
 
-    public int MaxHealth { get { return 1; } }
+    public float MaxHealth { get { return 1; } }
 
     public bool IsWeakPoint { get; } = false;
 
@@ -102,7 +102,7 @@ public class ActionBox : MonoBehaviour, IDamageable
             ThisActivate();
     }
     private bool wasHit => CurrentHealth == 0;
-    public void TakeDamage(int Damage, Guns.GunType gunType)
+    public void TakeDamage(float Damage, Guns.GunType gunType)
     {
         if(wasHit)
             return;
